+++ src/awscdk/index.ts
@@ -1,6 +1,8 @@
-export * from './java-app';
+export * from './auto-discover';
+export * from './awscdk-app-ts';
+export * from './awscdk-construct';
 export * from './cdk-config';
 export * from './cdk-tasks';
-export * from './auto-discover';
-export * from './lambda-function';
 export * from './integration-test';
+export * from './java-app';
+export * from './lambda-function';
